SimulationAnalysis.jl

SimulationAnalysis.KSpace — Type

KSpace{dims, S, OA}

A struct to hold the k-space information of a simulation.

Fields

s::S: The simulation object.

Nk::Int64: The number of k-vectors.

k_lengths::Array{Float64, 1}: The lengths of the k-vectors.

k_array::Array{Float64, 2}: The k-vectors.

kfactor::Int64: The k-factor.

cartesian_to_linear::OA: A mapping from Cartesian to linear indices.

SimulationAnalysis.MultiComponentDensityModes — Type

MultiComponentDensityModes

A struct to hold the density modes of a multi-component simulation.

Fields

Re::Vector{Array{Float64, 2}}: The real part of the density modes for each species.

Im::Vector{Array{Float64, 2}}: The imaginary part of the density modes for each species.

SimulationAnalysis.MultiComponentSimulation — Type

MultiComponentSimulation

A struct to hold the data of a multi-component simulation.

Fields

N::Int64: Total number of particles.

Ndims::Int64: Number of dimensions.

N_species::Int64: Number of species.

N_particles_per_species::Vector{Int}: Number of particles per species.

r_array::Vector{Array{Float64, 3}}: Particle positions for each species.

v_array::Vector{Array{Float64, 3}}: Particle velocities for each species.

F_array::Vector{Array{Float64, 3}}: Particle forces for each species.

t_array::Vector{Float64}: Time steps.

box_sizes::Vector{Float64}: Box sizes.

dt_array::Vector{Int}: Time step differences.

t1_t2_pair_array::Vector{Array{Int64, 2}}: Time pairs.

filepath::String: Path to the simulation file.

SimulationAnalysis.SingleComponentDensityModes — Type

SingleComponentDensityModes

A struct to hold the density modes of a single-component simulation.

Fields

Re::Array{Float64, 2}: The real part of the density modes.

Im::Array{Float64, 2}: The imaginary part of the density modes.

SimulationAnalysis.SingleComponentSimulation — Type

SingleComponentSimulation

A struct to hold the data of a single-component simulation.

Fields

N::Int64: Number of particles.

Ndims::Int64: Number of dimensions.

r_array::Array{Float64, 3}: Particle positions.

v_array::Array{Float64, 3}: Particle velocities.

F_array::Array{Float64, 3}: Particle forces.

D_array::Array{Float64, 1}: Diffusion coefficient.

t_array::Array{Float64, 1}: Time steps.

box_sizes::Vector{Float64}: Box sizes.

dt_array::Array{Int64, 1}: Time step differences.

t1_t2_pair_array::Vector{Array{Int64, 2}}: Time pairs.

filepath::String: Path to the simulation file.

SimulationAnalysis._find_density_modes! — Method

_find_density_modes!(Reρ, Imρ, r, kspace)

Calculate the density modes for a given simulation r and wave vector space kspace.

Reρ and Imρ arrays are preallocated arrays with size (N_timesteps, Nk) for each species of the simulation.

Arguments

Reρ: preallocated array to store the real part of the density modes.

Imρ: preallocated array to store the imaginary part of the density modes.

r: array containing the positions of the particles at each time step. It has shape (Ndim, N, Ntimesteps), where Ndim is the number of dimensions, N is the number of particles and Ntimesteps is the number of time steps.

kspace: the wave vector space used to calculate the density modes.

Returns

Nothing: The Reρ and Imρ arrays are updated in-place.

SimulationAnalysis.calculate_forces! — Method

calculate_forces!(s, forcestype::Bool)

Calculate the forces for a simulation.

This is a placeholder function that throws an error if forcestype is true.

Arguments

s: The simulation object.

forcestype::Bool: Whether to calculate forces.

SimulationAnalysis.calculate_forces! — Method

calculate_forces!(s, U::KAWCA; cutoff=2.0^(1.0/6.0))

Recalculates the total force on all particles in a multi-component simulation 's' according to the Kob-Andersen Weeks-Chandler-Andersen system U, and updates the force array in 's' with the calculated values.

Arguments

s: a MultiComponentSimulation object that contains the position and force arrays for each particle.

U::KAWCA: the interaction type.

cutoff: the cutoff distance for the potential energy function. Default is 2.0^(1.0/6.0).

SimulationAnalysis.construct_k_space — Method

construct_k_space(s::Simulation, bounds; kfactor=1, negative=false, rectangular=false)

Construct the k-space of a simulation.

Arguments

s::Simulation: The simulation object.

bounds: The bounds of the k-space.

kfactor=1: The k-factor.

negative=false: Whether to include negative k-vectors.

rectangular=false: Whether to use a rectangular k-space.

Returns

A KSpace object.

SimulationAnalysis.find_allowed_t1_t2_pair_array_log_multstarts — Method

find_allowed_t1_t2_pair_array_log_multstarts(t_array, N_starts)

Find allowed time pairs for a logarithmic time array with multiple starts.

Arguments

t_array: The time array.

N_starts: The number of starting points.

Returns

A tuple containing the dt_array and t1_t2_pair_array.

SimulationAnalysis.find_allowed_t1_t2_pair_array_quasilog — Method

find_allowed_t1_t2_pair_array_quasilog(t_array; doublefactor=150)

Find allowed time pairs for a quasi-logarithmic time array.

Arguments

t_array: The time array.

doublefactor=150: The factor by which to double the time step.

Returns

A tuple containing the dt_array and t1_t2_pair_array.

SimulationAnalysis.find_correlation_function! — Method

find_correlation_function!(AB, A, B, kspace::KSpace, dt_array, t1_t2_pair_array)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is averaged over the first dimension of A and B, a given set of time differences, and a given set of pairs of times. The correlation function is computed for each time difference and is stored in the array AB.

Arguments

AB::Array{Complex{Float64}, 1}: The array where the correlation function is stored.

A::Array{Complex{Float64}, 2}: The first list of observables. The first index is the wavevector index and the second index is the time index.

B::Array{Complex{Float64}, 2}: The second list of observables.

dt_array::Array{Float64, 1}: The array of time differences.

t1_t2_pair_array::Array{Array{Int64, 2}, 1}: The array of pairs of times. Each element of the array is a matrix with two columns. Each row of the matrix contains the indices of the times t1 and t2 for which the correlation function is computed.

Returns

Nothing: The correlation function is stored in the array AB.

SimulationAnalysis.find_correlation_function — Method

find_correlation_function(A, B, s)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is averaged over the first dimension of A and B.

Arguments

A::Array{Complex{Float64}, 2}: The first list of observables.

B::Array{Complex{Float64}, 2}: The second list of observables.

s::Simulation: The simulation.

Returns

C::Array{Complex{Float64}, 1}: The correlation function.

SimulationAnalysis.find_correlation_matrix — Method

find_correlation_matrix(observables_list1, observables_list2, kspace, s)

Computes the correlation matrix between two lists of observables. The observables are assumed to be complex-valued. The correlation matrix is a matrix of complex numbers.

Arguments

observables_list1::Array{Array{Complex{Float64}, 2}, 1}: The first list of observables.

observables_list2::Array{Array{Complex{Float64}, 2}, 1}: The second list of observables.

s::Simulation: The simulation.

Returns

matrix::Array{Array{Complex{Float64}, 1}, 2}: The correlation matrix.

SimulationAnalysis.find_density_modes — Method

find_density_modes(s::MultiComponentSimulation, kspace::KSpace; verbose=true)

Calculates the density modes for a multi-component simulation.

Arguments

s::MultiComponentSimulation: The simulation.

kspace::KSpace: The k-space.

verbose::Bool=true: Whether to print verbose output.

Returns

MultiComponentDensityModes: The density modes.

SimulationAnalysis.find_density_modes — Method

find_density_modes(s::SingleComponentSimulation, kspace::KSpace; verbose=true)

Calculates the density modes for a single-component simulation.

Arguments

s::SingleComponentSimulation: The simulation.

kspace::KSpace: The k-space.

verbose::Bool=true: Whether to print verbose output.

Returns

SingleComponentDensityModes: The density modes.

SimulationAnalysis.find_diagonal_four_point_k_vector_set — Method

find_diagonal_four_point_k_vector_set(kspace::KSpace3D, k1_bounds, k2_bounds, cosθ_bounds)    

Finds the set of k vectors that satisfy the conditions given for k1, k2 and cosθ12.
These are specified in 2-Tuples.

SimulationAnalysis.find_k_array — Method

find_k_array(bounds, box_sizes, dims; kfactor=1, rectangular=false, negative=false)

Find the k-vectors in a given k-space.

Arguments

bounds: The bounds of the k-space.

box_sizes: The box sizes.

dims: The number of dimensions.

kfactor=1: The k-factor.

rectangular=false: Whether to use a rectangular k-space.

negative=false: Whether to include negative k-vectors.

Returns

A tuple containing the k_lengths, k_array, and cartesian_to_linear.

SimulationAnalysis.find_k_array_2D — Method

find_k_array_2D(bounds, box_sizes; kfactor=1, rectangular=false, negative=false)

Find the k-vectors in a 2D k-space.

Arguments

bounds: The bounds of the k-space.

box_sizes: The box sizes.

kfactor=1: The k-factor.

rectangular=false: Whether to use a rectangular k-space.

negative=false: Whether to include negative k-vectors.

Returns

A tuple containing the k_lengths, k_array, and cartesian_to_linear.

SimulationAnalysis.find_k_array_3D — Method

find_k_array_3D(bounds, box_sizes; kfactor=1, rectangular=false, negative=false)

Find the k-vectors in a 3D k-space.

Arguments

bounds: The bounds of the k-space.

box_sizes: The box sizes.

kfactor=1: The k-factor.

rectangular=false: Whether to use a rectangular k-space.

negative=false: Whether to include negative k-vectors.

Returns

A tuple containing the k_lengths, k_array, and cartesian_to_linear.

SimulationAnalysis.find_log_time_array_multiple_starts — Method

find_log_time_array_multiple_starts(log_factor, N_starts, N_max)

Create a logarithmic time array with multiple starts.

Arguments

log_factor: The logarithmic factor.

N_starts: The number of starting points.

N_max: The maximum number of steps.

Returns

A logarithmic time array with multiple starts.

SimulationAnalysis.find_mean_squared_displacement — Method

Calculates the mean squared displacement from the t1, t2 pairs. Returns: An array contaning the mean squared displacement An array contaning the mean squared displacement per particle

SimulationAnalysis.find_mobile_clusters — Method

find_mobile_clusters(mobile_particles, neighbourlists)

Find clusters of mobile particles in a system. A cluster is defined as a set of particles that are connected by a bond. The function returns a vector of sets, where each set contains the indices of the particles in a cluster.

Arguments

mobile_particles::Set{Int}: The indices of the mobile particles.

neighbourlists::Vector{Vector{Int}}: The neighbourlists of the particles.

Returns

clusters::Vector{Set{Int}}: The clusters of mobile particles.

SimulationAnalysis.find_original_trajectories — Method

find_original_trajectories(r, box_sizes)

Reconstruct the original trajectories from the remapped ones.

Arguments

r: The remapped trajectories.

box_sizes: The box sizes.

Returns

The original trajectories.

SimulationAnalysis.find_quasilog_time_array — Method

find_quasilog_time_array(maxsteps; doublefactor=10)

Create a quasi-logarithmic time array.

Arguments

maxsteps: The maximum number of steps.

doublefactor=10: The factor by which to double the time step.

Returns

A quasi-logarithmic time array.

SimulationAnalysis.find_structure_factor — Method

Documentation for function findstructurefactor:

This function computes the static structure factor for a given simulation s using the density modes in ρkt and the k-space information in kspace. The structure factor is calculated separately for each species combination.

Arguments:

s: A MultiComponentSimulation object representing the simulation data.

kspace: A KSpace object containing information about the k-space grid.

ρkt: A MultiComponentDensityModes object representing the Fourier transform of the density fields for each species.

kmin: The minimum k value to include in the calculation. Default value is 0.0.

kmax: The maximum k value to include in the calculation. Default value is 10^10.

Returns:

Sk: A square matrix of size Nspecies x Nspecies representing the static structure factor. Sk[i, j] represents the structure factor for species i and j.

The function iterates over all possible species combinations (α, β), and calls the realstaticcorrelation_function function to calculate the structure factor for each combination. The result is stored in the Sk matrix. The function then returns Sk normalized by the total number of particles in the simulation.

SimulationAnalysis.find_voronoi_neighborlists — Method

find_neighborlists_voronoi(s::Simulation; max_distance_from_boundary=3.0) where Ndims

Computes the Voronoi neighborlists for a simulation.

Arguments

s::Simulation: The simulation.

max_distance_from_boundary::Float64: The maximum distance from the boundary to take periodic images into account.

Returns

neighbourlists::Vector{Vector{Vector{Int}}}: The Voronoi neighborlists.

SimulationAnalysis.read_Brownian_KALJ_simulation — Method

read_Brownian_KALJ_simulation(filenamefull, dt; maxt=-1, every=1, original=false, forces=true)

Read a Brownian simulation of a Kob-Andersen Lennard-Jones mixture. The file was generated with LAMMPS. The columns in the data file correspond to the following:

Atom ID

Atom type

x position

y position

z position

Arguments

filenamefull::String: Path to the simulation file.

dt::Float64: Time step.

maxt::Int=-1: Maximum number of time steps to read.

every::Int=1: Read every every-th time step.

original::Bool=false: Whether to reconstruct the original trajectories.

forces::Bool=true: Whether to compute forces.

Returns

MultiComponentSimulation: A MultiComponentSimulation object.

SimulationAnalysis.read_Brownian_KAWCA_simulation — Method

read_Brownian_KAWCA_simulation(filenamefull, dt; maxt=-1, every=1, original=false, forces=true)

Read a Brownian simulation of a Kob-Andersen WCA mixture. The file was generated with LAMMPS. The columns in the data file correspond to the following:

Atom ID

Atom type

x position

y position

z position

Arguments

filenamefull::String: Path to the simulation file.

dt::Float64: Time step.

maxt::Int=-1: Maximum number of time steps to read.

every::Int=1: Read every every-th time step.

original::Bool=false: Whether to reconstruct the original trajectories.

forces::Bool=true: Whether to compute forces.

Returns

MultiComponentSimulation: A MultiComponentSimulation object.

SimulationAnalysis.read_Newtonian_KAWCA_simulation — Method

read_Newtonian_KAWCA_simulation(filenamefull, dt; maxt=-1, every=1, original=false)

Read a Newtonian simulation of a Kob-Andersen WCA mixture. The file was generated with LAMMPS. The columns in the data file correspond to the following:

Atom ID

Atom type

x position

y position

z position

x velocities

y velocities

z velocities

Arguments

filenamefull::String: Path to the simulation file.

dt::Float64: Time step.

maxt::Int=-1: Maximum number of time steps to read.

every::Int=1: Read every every-th time step.

original::Bool=false: Whether to reconstruct the original trajectories.

Returns

MultiComponentSimulation: A MultiComponentSimulation object.

SimulationAnalysis.read_WCA_simulation — Method

read_WCA_simulation(filenamefull, dt; maxt=-1, every=1, original=false)

Read a simulation of WCA particles done with LAMMPS. in the data file, the columns correspond to the following:

Atom ID

x position

y position

z position

x force

y force

z force

Arguments

filenamefull::String: Path to the simulation file.

dt::Float64: Time step.

maxt::Int=-1: Maximum number of time steps to read.

every::Int=1: Read every every-th time step.

original::Bool=false: Whether to reconstruct the original trajectories.

Returns

SingleComponentSimulation: A SingleComponentSimulation object.

SimulationAnalysis.read_continuously_hard_sphere_simulation — Method

read_continuously_hard_sphere_simulation(filename; original=false, velocities=false, forcestype=false, time_origins="quasilog")

Read a simulation of continuously polydisperse particles. The file was generated with SimulationCode.jl.

Arguments

filename::String: Path to the simulation file.

original::Bool=false: Whether to reconstruct the original trajectories.

velocities::Bool=false: Whether to read velocities.

forcestype=false: Type of forces to compute.

time_origins="quasilog": How to select the time origins.

Returns

SingleComponentSimulation: A SingleComponentSimulation object.

SimulationAnalysis.read_monodisperse_hard_sphere_simulation — Method

read_monodisperse_hard_sphere_simulation(filename; original=false, velocities=false, forcestype=false, dtarr=true)

Read a simulation of monodisperse particles. The file was generated with SimulationCode.jl.

Arguments

filename::String: Path to the simulation file.

original::Bool=false: Whether to reconstruct the original trajectories.

velocities::Bool=false: Whether to read velocities.

forcestype=false: Type of forces to compute.

dtarr::Bool=true: Whether to compute the time step array.

Returns

SingleComponentSimulation: A SingleComponentSimulation object.

SimulationAnalysis.real_correlation_function! — Method

real_correlation_function!(C, ReA, ImA, ReB, ImB, kspace, dt_array, t1_t2_pair_array, kmin, kmax)

Computes the real part of the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is computed for a given k-space, a given set of time differences, and a given set of pairs of times. The correlation function is computed for each time difference and is stored in the array C.

Arguments

C::Array{Float64, 1}: The array where the correlation function is stored.

ReA::Array{Float64, 2}: The real part of the first list of observables. The first index is the time index and the second index is the wavevector index.

ImA::Array{Float64, 2}: The imaginary part of the first list of observables.

ReB::Array{Float64, 2}: The real part of the second list of observables.

ImB::Array{Float64, 2}: The imaginary part of the second list of observables.

kspace::KSpace: The k-space.

dt_array::Array{Float64, 1}: The array of time differences.

t1_t2_pair_array::Array{Array{Int64, 2}, 1}: The array of pairs of times. Each element of the array is a matrix with two columns. Each row of the matrix contains the indices of the times t1 and t2 for which the correlation function is computed.

kmin::Float64: The minimum value of the wavevector.

kmax::Float64: The maximum value of the wavevector.

Returns

Nothing: The correlation function is stored in the array C.

SimulationAnalysis.real_static_correlation_function — Method

real_static_correlation_function(ReA::Array{Float64, 2}, ImA::Array{Float64, 2}, ReB::Array{Float64, 2}, ImB::Array{Float64, 2}, kspace::KSpace, kmin, kmax)

Computes the real-valued static correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a real number.

Arguments

ReA::Array{Float64, 2}: The real part of the first list of observables.

ImA::Array{Float64, 2}: The imaginary part of the first list of observables.

ReB::Array{Float64, 2}: The real part of the second list of observables.

ImB::Array{Float64, 2}: The imaginary part of the second list of observables.

kspace::KSpace: The k-space.

kmin::Float64: The minimum value of the wavevector.

kmax::Float64: The maximum value of the wavevector.

Returns

C::Float64: The correlation function.

SimulationAnalysis.remap_positions! — Method

remap_positions!(r::Array{Float64,3}, box_sizes)

Remap the positions to be inside the box.

Arguments

r::Array{Float64,3}: The particle positions.

box_sizes: The box sizes.

SimulationAnalysis.reshape_data — Method

reshape_data(r::Array{Array{Float64,2},1})

Reshape the data from a vector of 2D arrays to a 3D array.

Arguments

r::Array{Array{Float64,2},1}: A vector of 2D arrays.

Returns

Array{Float64,3}: A 3D array.

SimulationAnalysis.separate_trajectories — Method

separate_trajectories(r, type_list)

Separate the trajectories of different species.

Arguments

r: The trajectories.

type_list: A list of particle types.

Returns

A vector of trajectories, one for each species.