API Reference

Simulation

SimulationAnalysis.SingleComponentSimulation — Type

SingleComponentSimulation

A struct to hold the data of a single-component simulation.

Fields

N::Int64: Number of particles.

Ndims::Int64: Number of dimensions.

r_array::Array{Float64, 3}: Particle positions.

v_array::Array{Float64, 3}: Particle velocities.

F_array::Array{Float64, 3}: Particle forces.

D_array::Array{Float64, 1}: Diffusion coefficient.

t_array::Array{Float64, 1}: Time steps.

box_sizes::Vector{Float64}: Box sizes.

dt_array::Array{Int64, 1}: Time step differences.

t1_t2_pair_array::Vector{Array{Int64, 2}}: Time pairs.

filepath::String: Path to the simulation file.

SimulationAnalysis.MultiComponentSimulation — Type

MultiComponentSimulation

A struct to hold the data of a multi-component simulation.

Fields

N::Int64: Total number of particles.

Ndims::Int64: Number of dimensions.

N_species::Int64: Number of species.

N_particles_per_species::Vector{Int}: Number of particles per species.

r_array::Vector{Array{Float64, 3}}: Particle positions for each species.

v_array::Vector{Array{Float64, 3}}: Particle velocities for each species.

F_array::Vector{Array{Float64, 3}}: Particle forces for each species.

t_array::Vector{Float64}: Time steps.

box_sizes::Vector{Float64}: Box sizes.

dt_array::Vector{Int}: Time step differences.

t1_t2_pair_array::Vector{Array{Int64, 2}}: Time pairs.

filepath::String: Path to the simulation file.

K-Space

SimulationAnalysis.KSpace — Type

KSpace{dims, S, OA}

A struct to hold the k-space information of a simulation.

Fields

s::S: The simulation object.

Nk::Int64: The number of k-vectors.

k_lengths::Array{Float64, 1}: The lengths of the k-vectors.

k_array::Array{Float64, 2}: The k-vectors.

kfactor::Int64: The k-factor.

cartesian_to_linear::OA: A mapping from Cartesian to linear indices.

SimulationAnalysis.construct_k_space — Function

construct_k_space(s::Simulation, bounds; kfactor=1, negative=false, rectangular=false)

Construct the k-space of a simulation.

Arguments

s::Simulation: The simulation object.

bounds: The bounds of the k-space.

kfactor=1: The k-factor.

negative=false: Whether to include negative k-vectors.

rectangular=false: Whether to use a rectangular k-space.

Returns

A KSpace object.

Correlation Function

SimulationAnalysis.find_correlation_function — Function

find_correlation_function(A, B, s)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is averaged over the first dimension of A and B.

Arguments

A::Array{Complex{Float64}, 2}: The first list of observables.

B::Array{Complex{Float64}, 2}: The second list of observables.

s::Simulation: The simulation.

Returns

C::Array{Complex{Float64}, 1}: The correlation function.

SimulationAnalysis.find_correlation_matrix — Function

find_correlation_matrix(observables_list1, observables_list2, kspace, s)

Computes the correlation matrix between two lists of observables. The observables are assumed to be complex-valued. The correlation matrix is a matrix of complex numbers.

Arguments

observables_list1::Array{Array{Complex{Float64}, 2}, 1}: The first list of observables.

observables_list2::Array{Array{Complex{Float64}, 2}, 1}: The second list of observables.

s::Simulation: The simulation.

Returns

matrix::Array{Array{Complex{Float64}, 1}, 2}: The correlation matrix.

Density Modes

SimulationAnalysis.find_density_modes — Function

find_density_modes(s::SingleComponentSimulation, kspace::KSpace; verbose=true)

Calculates the density modes for a single-component simulation.

Arguments

s::SingleComponentSimulation: The simulation.

kspace::KSpace: The k-space.

verbose::Bool=true: Whether to print verbose output.

Returns

SingleComponentDensityModes: The density modes.

find_density_modes(s::MultiComponentSimulation, kspace::KSpace; verbose=true)

Calculates the density modes for a multi-component simulation.

Arguments

s::MultiComponentSimulation: The simulation.

kspace::KSpace: The k-space.

verbose::Bool=true: Whether to print verbose output.

Returns

MultiComponentDensityModes: The density modes.

SimulationAnalysis.SingleComponentDensityModes — Type

SingleComponentDensityModes

A struct to hold the density modes of a single-component simulation.

Fields

Re::Array{Float64, 2}: The real part of the density modes.

Im::Array{Float64, 2}: The imaginary part of the density modes.

SimulationAnalysis.MultiComponentDensityModes — Type

MultiComponentDensityModes

A struct to hold the density modes of a multi-component simulation.

Fields

Re::Vector{Array{Float64, 2}}: The real part of the density modes for each species.

Im::Vector{Array{Float64, 2}}: The imaginary part of the density modes for each species.

Forces

SimulationAnalysis.Weysser — Type

Weysser <: InteractionPotential

A struct for the Weysser interaction potential.

Fields

ϵ::Float64: The interaction strength.

δ::Float64: The interaction range.

SimulationAnalysis.Berthier — Type

Berthier <: InteractionPotential

A struct for the Berthier interaction potential.

Fields

c0::Float64: The c0 coefficient.

c2::Float64: The c2 coefficient.

c4::Float64: The c4 coefficient.

ζ::Float64: The zeta coefficient.

σ_ratio::Float64: The ratio of the diameters.

SimulationAnalysis.WCA — Type

WCA <: InteractionPotential

A struct for the WCA interaction potential.

Fields

ϵ::Float64: The interaction strength.

σ::Float64: The interaction range.

SimulationAnalysis.KAWCA — Type

KAWCA <: InteractionPotential

A struct for the Kob-Andersen WCA interaction potential.

Fields

ϵ11::Float64: The interaction strength between particles of type 1.

ϵ12::Float64: The interaction strength between particles of type 1 and 2.

ϵ22::Float64: The interaction strength between particles of type 2.

σ11::Float64: The interaction range between particles of type 1.

σ12::Float64: The interaction range between particles of type 1 and 2.

σ22::Float64: The interaction range between particles of type 2.

SimulationAnalysis.calculate_forces! — Function

calculate_forces!(s, forcestype::Bool)

Calculate the forces for a simulation.

This is a placeholder function that throws an error if forcestype is true.

Arguments

s: The simulation object.

forcestype::Bool: Whether to calculate forces.

calculate_forces!(s, U; friction=false, cutoff=1.25)

Recalculates the total force on all particles according to the langevin equation F = -∇U - γv. This function also updates the total potential energy in the output datastructure.

calculate_forces!(s, U::KAWCA; cutoff=2.0^(1.0/6.0))

Recalculates the total force on all particles in a multi-component simulation s according to the Kob-Andersen Weeks-Chandler-Andersen system U, and updates the force array in s with the calculated values.

Arguments

s: a MultiComponentSimulation object that contains the position and force arrays for each particle.

U::KAWCA: the interaction type.

cutoff: the cutoff distance for the potential energy function. Default is 2.0^(1.0/6.0).

Intermediate Scattering Function

SimulationAnalysis.find_intermediate_scattering_function — Function

find_intermediate_scattering_function(s::Simulation; kmin=7.0, kmax=7.4, kfactor=1)

Calculates the intermediate scattering function for a simulation.

Arguments

s::Simulation: The simulation.

kmin=7.0: The minimum k value.

kmax=7.4: The maximum k value.

kfactor=1: The k-factor.

Returns

F: The intermediate scattering function.

find_intermediate_scattering_function(s::Simulation, kspace::KSpace, ρkt, k_sample_array::AbstractVector; k_binwidth=0.1)

Calculates the intermediate scattering function for a simulation for a given array of k-values.

Arguments

s::Simulation: The simulation.

kspace::KSpace: The k-space.

ρkt: The density modes.

k_sample_array::AbstractVector: The array of k-values.

k_binwidth=0.1: The bin width for k.

Returns

F_array: An array of intermediate scattering functions.

find_intermediate_scattering_function(s::SingleComponentSimulation, kspace::KSpace, ρkt::SingleComponentDensityModes; kmin=0.0, kmax=10.0^10.0)

Calculates the intermediate scattering function for a single-component simulation.

Arguments

s::SingleComponentSimulation: The simulation.

kspace::KSpace: The k-space.

ρkt::SingleComponentDensityModes: The density modes.

kmin=0.0: The minimum k value.

kmax=10.0^10.0: The maximum k value.

Returns

Fk: The intermediate scattering function.

find_intermediate_scattering_function(s::MultiComponentSimulation, kspace::KSpace, ρkt::MultiComponentDensityModes; kmin=0.0, kmax=10.0^10.0)

Calculates the intermediate scattering function for a multi-component simulation.

Arguments

s::MultiComponentSimulation: The simulation.

kspace::KSpace: The k-space.

ρkt::MultiComponentDensityModes: The density modes.

kmin=0.0: The minimum k value.

kmax=10.0^10.0: The maximum k value.

Returns

Fk: The intermediate scattering function.

SimulationAnalysis.find_self_intermediate_scattering_function — Function

find_self_intermediate_scattering_function(s::Simulation; kmin=7.0, kmax=7.4, kfactor=1)

Calculates the self-intermediate scattering function for a simulation.

Arguments

s::Simulation: The simulation.

kmin=7.0: The minimum k value.

kmax=7.4: The maximum k value.

kfactor=1: The k-factor.

Returns

F: The self-intermediate scattering function.

find_self_intermediate_scattering_function(s::Simulation, kspace::KSpace, k_sample_array::AbstractVector; k_binwidth=0.1)

Calculates the self-intermediate scattering function for a simulation for a given array of k-values.

Arguments

s::Simulation: The simulation.

kspace::KSpace: The k-space.

k_sample_array::AbstractVector: The array of k-values.

k_binwidth=0.1: The bin width for k.

Returns

F_array: An array of self-intermediate scattering functions.

find_self_intermediate_scattering_function(s::SingleComponentSimulation, kspace::KSpace; kmin=0.0, kmax=10.0^10.0)

Calculates the self-intermediate scattering function for a single-component simulation.

Arguments

s::SingleComponentSimulation: The simulation.

kspace::KSpace: The k-space.

kmin=0.0: The minimum k value.

kmax=10.0^10.0: The maximum k value.

Returns

Fk: The self-intermediate scattering function.

Fks_per_particle: The self-intermediate scattering function per particle.

Structure Factors

SimulationAnalysis.find_structure_factor — Function

find_structure_factor(s::Simulation; kmin=7.0, kmax=7.4, kfactor=1)

Calculates the structure factor for a simulation.

Arguments

s::Simulation: The simulation.

kmin=7.0: The minimum k value.

kmax=7.4: The maximum k value.

kfactor=1: The k-factor.

Returns

S: The structure factor.

find_structure_factor(s::Simulation, kspace::KSpace, ρkt::AbstractDensityModes, k_sample_array::AbstractVector; k_binwidth=0.1)

Calculates the structure factor for a simulation for a given array of k-values.

Arguments

s::Simulation: The simulation.

kspace::KSpace: The k-space.

ρkt::AbstractDensityModes: The density modes.

k_sample_array::AbstractVector: The array of k-values.

k_binwidth=0.1: The bin width for k.

Returns

S_array: An array of structure factors.

find_structure_factor(s::SingleComponentSimulation, kspace::KSpace, ρkt::SingleComponentDensityModes; kmin=0.0, kmax=10.0^10.0)

Calculates the structure factor for a single-component simulation.

Arguments

s::SingleComponentSimulation: The simulation.

kspace::KSpace: The k-space.

ρkt::SingleComponentDensityModes: The density modes.

kmin=0.0: The minimum k value.

kmax=10.0^10.0: The maximum k value.

Returns

Sk: The structure factor.

find_structure_factor(s::MultiComponentSimulation, kspace::KSpace, ρkt::MultiComponentDensityModes; kmin=0.0, kmax=10.0^10.0)

This function computes the static structure factor for a given simulation s using the density modes in ρkt and the k-space information in kspace. The structure factor is calculated separately for each species combination.

Arguments

s: A MultiComponentSimulation object representing the simulation data.

kspace: A KSpace object containing information about the k-space grid.

ρkt: A MultiComponentDensityModes object representing the Fourier transform of the density fields for each species.

kmin: The minimum k value to include in the calculation. Default value is 0.0.

kmax: The maximum k value to include in the calculation. Default value is 10^10.

Returns

Sk: A square matrix of size N_species x N_species representing the static structure factor. Sk[i, j] represents the structure factor for species i and j.

Mean Squared Displacement

SimulationAnalysis.find_mean_squared_displacement — Function

find_mean_squared_displacement(simulation::Simulation; per_particle=false)

Calculates the mean squared displacement for a simulation.

Arguments

simulation::Simulation: The simulation.

per_particle=false: Whether to return the mean squared displacement per particle.

Returns

msd: The mean squared displacement.

msd_per_particle: The mean squared displacement per particle (if per_particle is true).

find_mean_squared_displacement(r, dt_array, t1_t2_pair_array; verbose=true)

Calculates the mean squared displacement from the t1, t2 pairs.

Arguments

r: The positions of the particles.

dt_array: The array of time differences.

t1_t2_pair_array: The array of pairs of times.

verbose=true: Whether to print verbose output.

Returns

msd: The mean squared displacement.

msd_per_particle: The mean squared displacement per particle.

Overlap Function

SimulationAnalysis.find_overlap_function — Function

find_overlap_function(s; a=0.5)

Calculates the overlap function for a simulation.

Arguments

s: The simulation.

a=0.5: The cutoff distance.

Returns

Fs: The overlap function.

Fs_pp: The overlap function per particle.

Neighborlists

SimulationAnalysis.find_relative_distance_neighborlists — Function

find_relative_distance_neighborlists(s, rc; ζ = 0.2)

Finds the neighbor lists for a simulation based on the relative distance between particles.

Arguments

s: The simulation.

rc: The cutoff distance.

ζ=0.2: The polydispersity parameter.

Returns

neighbourlists: The neighbor lists.

SimulationAnalysis.find_absolute_distance_neighborlists — Function

find_absolute_distance_neighborlists(s, rc)

Finds the neighbor lists for a simulation based on the absolute distance between particles.

Arguments

s: The simulation.

rc: The cutoff distance.

Returns

neighbourlists: The neighbor lists.

SimulationAnalysis.find_voronoi_neighborlists — Function

find_voronoi_neighborlists(s; max_distance_from_boundary=3.0, verbose=true, indices=eachindex(s.t_array))

Computes the Voronoi neighborlists for a simulation.

Arguments

s: The simulation.

max_distance_from_boundary=3.0: The maximum distance from the boundary to take periodic images into account.

verbose=true: Whether to print verbose output.

indices=eachindex(s.t_array): The indices of the time steps to consider.

Returns

neighbourlists: The Voronoi neighborlists.

Bond Breaking Parameter

SimulationAnalysis.find_CB — Function

find_CB(s, neighbourlists1, neighbourlists2)

Calculates the bond-breaking parameter CB for all particles.

Utils

SimulationAnalysis.find_relaxation_time — Function

find_relaxation_time(t, F; threshold=exp(-1))

Calculates the relaxation time of a function F(t).

The relaxation time is defined as the time at which F(t) drops below a certain threshold. The function uses linear interpolation on a logarithmic time scale to find the relaxation time.

Arguments

t: The time points.

F: The function values.

threshold=exp(-1): The threshold.

Returns

The relaxation time.

Radial Distribution Function

SimulationAnalysis.find_radial_distribution_function — Function

find_radial_distribution_function(s::MultiComponentSimulation, Nbins, rmax)

Calculates the radial distribution function for a multi-component simulation.

find_radial_distribution_function(s::SingleComponentSimulation, Nbins::Int, rmax; Ntasks=1)

Calculates the radial distribution function for a single-component simulation.

find_radial_distribution_function(s::SingleComponentSimulation, dr::Float64, rmax; Ntasks=1)

Calculates the radial distribution function for a single-component simulation.