SimulationAnalysis.jl

SimulationAnalysis._find_density_modes! — Method

Calculate the density modes for a given simulation r and wave vector space kspace.

Reρ and Imρ arrays are preallocated arrays with size (N_timesteps, Nk) for each species of the simulation.

Args: - Reρ (array): preallocated array to store the real part of the density modes. - Imρ (array): preallocated array to store the imaginary part of the density modes. - r (array): array containing the positions of the particles at each time step. It has shape (Ndim, N, Ntimesteps), where Ndim is the number of dimensions, N is the number of particles and Ntimesteps is the number of time steps. - kspace (KSpace): the wave vector space used to calculate the density modes.

Returns: - None: Reρ and Imρ arrays are updated in-place.

SimulationAnalysis.calculate_forces! — Method

calculate_forces!(s, U::KAWCA; cutoff=2.0^(1.0/6.0))

Recalculates the total force on all particles in a multi-component simulation 's' according to the Kob-Andersen Weeks-Chandler-Andersen system U, and updates the force array in 's' with the calculated values.

Arguments

s: a MultiComponentSimulation object that contains the position and force arrays for each particle.

U::KAWCA: the interaction type.

cutoff: the cutoff distance for the potential energy function. Default is 2.0^(1.0/6.0).

SimulationAnalysis.find_correlation_function! — Method

find_correlation_function!(AB, A, B, kspace::KSpace, dt_array, t1_t2_pair_array)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is averaged over the first dimension of A and B, a given set of time differences, and a given set of pairs of times. The correlation function is computed for each time difference and is stored in the array AB.

Arguments

AB::Array{Complex{Float64}, 1}: The array where the correlation function is stored.

A::Array{Complex{Float64}, 2}: The first list of observables. The first index is the wavevector index and the second index is the time index.

B::Array{Complex{Float64}, 2}: The second list of observables.

dt_array::Array{Float64, 1}: The array of time differences.

t1_t2_pair_array::Array{Array{Int64, 2}, 1}: The array of pairs of times. Each element of the array is a matrix with two columns. Each row of the matrix contains the indices of the times t1 and t2 for which the correlation function is computed.

Returns

Nothing: The correlation function is stored in the array AB.

SimulationAnalysis.find_correlation_function — Method

find_correlation_function(A, B, s)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is averaged over the first dimension of A and B.

Arguments

A::Array{Complex{Float64}, 2}: The first list of observables.

B::Array{Complex{Float64}, 2}: The second list of observables.

s::Simulation: The simulation.

Returns

C::Array{Complex{Float64}, 1}: The correlation function.

SimulationAnalysis.find_correlation_matrix — Method

find_correlation_matrix(observables_list1, observables_list2, kspace, s)

Computes the correlation matrix between two lists of observables. The observables are assumed to be complex-valued. The correlation matrix is a matrix of complex numbers.

Arguments

observables_list1::Array{Array{Complex{Float64}, 2}, 1}: The first list of observables.

observables_list2::Array{Array{Complex{Float64}, 2}, 1}: The second list of observables.

s::Simulation: The simulation.

Returns

matrix::Array{Array{Complex{Float64}, 1}, 2}: The correlation matrix.

SimulationAnalysis.find_density_modes — Method

finddensitymodes(s::MultiComponentSimulation, kspace::KSpace; verbose=true)

This function calculates the density modes for a multi-component simulation s using the wave vectors in kspace. The density modes are calculated for each species in the simulation and returned as a MultiComponentDensityModes object.

Arguments:

s::MultiComponentSimulation: A MultiComponentSimulation object representing the simulation.

kspace::KSpace: A KSpace object representing the wave vectors to use for calculating the density modes.

verbose::Bool=true: A flag indicating whether to print verbose output.

Returns:

modes::MultiComponentDensityModes: A MultiComponentDensityModes object containing the real and imaginary parts of the density modes for each species.

SimulationAnalysis.find_diagonal_four_point_k_vector_set — Method

find_diagonal_four_point_k_vector_set(kspace::KSpace3D, k1_bounds, k2_bounds, cosθ_bounds)    

Finds the set of k vectors that satisfy the conditions given for k1, k2 and cosθ12.
These are specified in 2-Tuples.

SimulationAnalysis.find_mean_squared_displacement — Method

Calculates the mean squared displacement from the t1, t2 pairs. Returns: An array contaning the mean squared displacement An array contaning the mean squared displacement per particle

SimulationAnalysis.find_mobile_clusters — Method

find_mobile_clusters(mobile_particles, neighbourlists)

Find clusters of mobile particles in a system. A cluster is defined as a set of particles that are connected by a bond. The function returns a vector of sets, where each set contains the indices of the particles in a cluster.

Arguments

mobile_particles::Set{Int}: The indices of the mobile particles.

neighbourlists::Vector{Vector{Int}}: The neighbourlists of the particles.

Returns

clusters::Vector{Set{Int}}: The clusters of mobile particles.

SimulationAnalysis.find_structure_factor — Method

Documentation for function findstructurefactor:

This function computes the static structure factor for a given simulation s using the density modes in ρkt and the k-space information in kspace. The structure factor is calculated separately for each species combination.

Arguments:

s: A MultiComponentSimulation object representing the simulation data.

kspace: A KSpace object containing information about the k-space grid.

ρkt: A MultiComponentDensityModes object representing the Fourier transform of the density fields for each species.

kmin: The minimum k value to include in the calculation. Default value is 0.0.

kmax: The maximum k value to include in the calculation. Default value is 10^10.

Returns:

Sk: A square matrix of size Nspecies x Nspecies representing the static structure factor. Sk[i, j] represents the structure factor for species i and j.

The function iterates over all possible species combinations (α, β), and calls the realstaticcorrelation_function function to calculate the structure factor for each combination. The result is stored in the Sk matrix. The function then returns Sk normalized by the total number of particles in the simulation.

SimulationAnalysis.find_voronoi_neighborlists — Method

find_neighborlists_voronoi(s::Simulation; max_distance_from_boundary=3.0) where Ndims

Computes the Voronoi neighborlists for a simulation.

Arguments

s::Simulation: The simulation.

max_distance_from_boundary::Float64: The maximum distance from the boundary to take periodic images into account.

Returns

neighbourlists::Vector{Vector{Vector{Int}}}: The Voronoi neighborlists.

SimulationAnalysis.real_correlation_function! — Method

find_correlation_function!(C, A, B, kspace, dt_array, t1_t2_pair_array)

Computes the correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a complex number. The correlation function is computed for a given k-space, a given set of time differences, and a given set of pairs of times. The correlation function is computed for each time difference and is stored in the array C.

Arguments

C::Array{Complex{Float64}, 1}: The array where the correlation function is stored.

A::Array{Complex{Float64}, 2}: The first list of observables. The first index is the wavevector index and the second index is the time index.

B::Array{Complex{Float64}, 2}: The second list of observables.

kspace::KSpace: The k-space.

dt_array::Array{Float64, 1}: The array of time differences.

t1_t2_pair_array::Array{Array{Int64, 2}, 1}: The array of pairs of times. Each element of the array is a matrix with two columns. Each row of the matrix contains the indices of the times t1 and t2 for which the correlation function is computed.

kmin::Float64: The minimum value of the wavevector.

kmax::Float64: The maximum value of the wavevector.

Returns

Nothing: The correlation function is stored in the array C.

SimulationAnalysis.real_static_correlation_function — Method

real_static_correlation_function(ReA::Array{Float64, 2}, ImA::Array{Float64, 2}, ReB::Array{Float64, 2}, ImB::Array{Float64, 2}, kspace::KSpace, kmin, kmax)

Computes the real-valued static correlation function between two lists of observables. The observables are assumed to be complex-valued. The correlation function is a real number.

Arguments

ReA::Array{Float64, 2}: The real part of the first list of observables.

ImA::Array{Float64, 2}: The imaginary part of the first list of observables.

ReB::Array{Float64, 2}: The real part of the second list of observables.

ImB::Array{Float64, 2}: The imaginary part of the second list of observables.

kspace::KSpace: The k-space.

kmin::Float64: The minimum value of the wavevector.

kmax::Float64: The maximum value of the wavevector.

Returns

C::Float64: The correlation function.